% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_na.R
\name{num_na}
\alias{num_na}
\title{counts number of na values of a variable in a df}
\usage{
num_na(df, var)
}
\arguments{
\item{df}{data frame or its extension}

\item{var}{unquoted column/variable name to count number of NA values}
}
\value{
a number of how many na values there contain in the input var
}
\description{
counts number of na values of a variable in a df
}
\examples{
num_na(mtcars, mpg)
num_na(mtcars, cyl)
num_na(mtcars, disp)
}
