% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fun.R
\name{summary_fun}
\alias{summary_fun}
\title{summary function that applies a function to whole data frame
while removing na values}
\usage{
summary_fun(df, fun)
}
\arguments{
\item{df}{A data frame that is interested to apply a summary
function to (eg. mean, standard deviation)}

\item{fun}{a function to apply for the input df}
}
\value{
a data.frame that shows a summary statistic for each column
in the original input df, and rounded to one decimal place
}
\description{
summary function that applies a function to whole data frame
while removing na values
}
\examples{
summary_fun(mtcars, mean)
summary_fun(mtcars, sd)
summary_fun(mtcars, max)
}
