% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{wrangle_data}
\alias{wrangle_data}
\title{Wrangle the df and keep only the interested variables (unquoted)
and output should have same type as of df.}
\usage{
wrangle_data(df, ...)
}
\arguments{
\item{df}{A data frame that needs to be wrangle}

\item{...}{unquoted variables names to select data}
}
\value{
A data frame that contains only those variables
specified
}
\description{
Wrangle the df and keep only the interested variables (unquoted)
and output should have same type as of df.
}
\examples{
wrangle_data(mtcars, mpg)
wrangle_data(mtcars, mpg, disp)
}
