% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_vars.R
\name{visualize_vars}
\alias{visualize_vars}
\title{Visualize a df with two specified variables and an optional alpha value}
\usage{
visualize_vars(df, x, y, alpha = 0.4)
}
\arguments{
\item{df}{A data frame or extension of it}

\item{x}{A numeric variable shown on x-axis}

\item{y}{A numeric variable shown on y-axis}

\item{alpha}{number is default to 0.4. If the viewers want to change anything here,
you can set alpha as a decimal number in the function.}
}
\value{
ggplot point graph of two variables from same df
}
\description{
Visualize a df with two specified variables and an optional alpha value
}
\examples{
visualize_vars(mtcars, mpg, cyl)
visualize_vars(data.frame(x = 1, y =2), x,y)
}
